Add skeleton code for activity components
